% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session5_HW_Function.R
\name{lindirect_utility}
\alias{lindirect_utility}
\title{An lindirect_utility function}
\usage{
lindirect_utility(x, y, px, py, a, b)
}
\description{
This function outpus the indirect utility for a log-linear utility (alog(X) + blog(Y))
}
\examples{
lindirect_utility()
}

